# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/collaboration/clubs/{clubId}/room": {
      "get": {
        "operationId": "Collaboration_getRoom",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get collaboration room state for a club",
        "tags": [
          "collaboration"
        ]
      }
    },
    "/api/v1/collaboration/clubs/{clubId}/messages": {
      "post": {
        "operationId": "Collaboration_createMessage",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollaborationMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Post a collaboration chat message",
        "tags": [
          "collaboration"
        ]
      }
    },
    "/api/v1/collaboration/clubs/{clubId}/tasks": {
      "post": {
        "operationId": "Collaboration_createTask",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollaborationTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a collaboration sprint task",
        "tags": [
          "collaboration"
        ]
      }
    },
    "/api/v1/collaboration/tasks/{taskId}": {
      "patch": {
        "operationId": "Collaboration_updateTask",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCollaborationTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update a collaboration sprint task",
        "tags": [
          "collaboration"
        ]
      }
    },
    "/api/v1/collaboration/clubs/{clubId}/code-submissions": {
      "post": {
        "operationId": "Collaboration_createCodeSubmission",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollaborationCodeSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Submit a collaboration code snippet",
        "tags": [
          "collaboration"
        ]
      }
    }
  },
  "info": {
    "title": "CodeCircle",
    "description": "API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Root"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for external services"
      }
    },
    "schemas": {
...
      "CreateCollaborationMessageDto": {
        "type": "object",
        "properties": {}
      },
      "CreateCollaborationTaskDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateCollaborationTaskDto": {
        "type": "object",
        "properties": {}
      },
      "CreateCollaborationCodeSubmissionDto": {
        "type": "object",
        "properties": {}
      }
    }
  },
...
}